Initial split of zebrad acceptance tests into modules
